requirements.txt
streamlit
pandas
numpy
joblib
google-genai
python-dotenv
folium
streamlit-folium
pydeck